// client/dist/widgets/common/list/src/config.ts
export interface ListItemData {
  columnIndex?: number
  rowIndex?: number
  previewDynamicStyle?: boolean

  hideCardTool: boolean
  selectionIsList: boolean
  selectionIsInList: boolean
  isEditing: boolean
  builderStatus: Status,
  changeIsResizingCard: (isResizingCard: boolean) => void

  dynamicStyleOfCard?: IMDynamicStyleOfCard
  useDataSources?: ImmutableArray<UseDataSource>

  index: number,
  isHover: boolean,
  record: DataRecord,
  active: boolean

  id: string
  widgetId: string
  config: IMConfig
  recordLength: number,
  widgetRect: ElementSize,
  currentCardSize: ElementSize,
  handleListMouseMove: (itemIndex: number) => void,
  handleListMouseLeave: () => void,

  browserSizeMode: BrowserSizeMode,
  isRTL: boolean,
  builderSupportModules: any,
  interact: any,
  appMode: AppMode,
  theme: IMThemeVariables,
  LayoutEntry: any,
  layouts: any,
  datasourceId: string,
  isWidthPercentage: boolean,
  formatMessage: any,
  selectCard: () => void,
  handleResizeCard: (newCardSize: any, resizeEnd?: boolean, isTop?: boolean, isLeft?: boolean, isReplace?: boolean) => void,
  onChange: (itemRecord: DataRecord) => void,
  updateCanClickLinkWhenClickItem: (itemRecord: DataRecord) => void,
  updateCardToolPosition: () => void
}

export interface DynamicStyleOfCard {
  [status: string]: DynamicStyle
}

export type IMDynamicStyleOfCard = ImmutableObject<DynamicStyleOfCard>

export enum SelectionModeType {
  None = 'NONE',
  Single = 'SINGLE',
  Multiple = 'MULTIPLE'
}

export enum PageStyle {
  Scroll = 'SCROLL',
  MultiPage = 'MULTIPAGE'
}

export enum AlignType {
  Start = 'FLEX-START',
  Center = 'CENTER',
  End = 'FLEX-END'
}

export enum DirectionType {
  Horizon = 'HORIZON',
  Vertical = 'VERTICAL'
}

export enum ListLayoutType {
  Row = 'ROW',
  Column = 'COLUMN',
  GRID = 'GRID'
}
// ...
export enum Status {
  Default = 'DEFAULT',
  Selected = 'SELECTED',
  Hover = 'HOVER'
}

export interface CardSize {
  height: number | string
  width: number | string
}

export interface ElementSize {
  height: number
  width: number
}
// ...
export interface Config {
  pageTransition?: PageTransitionType
  hoverType?: HoverType
  selectedStyle?: SelectedStyle
  differentOddEven?: boolean
  itemStyle?: ItemStyle
  isCheckEmptyTemplate?: boolean
  isItemStyleConfirm?: boolean
  direction?: DirectionType
  alignType?: AlignType
  space?: number
  horizontalSpace?: number
  verticalSpace?: number
  itemsPerPage?: number
  pageStyle?: PageStyle
  scrollBarOpen?: boolean
  navigatorOpen?: boolean
  scrollStep?: number
  style?: ImmutableObject<WidgetStyle>
  isInitialed?: boolean
  lockItemRatio?: boolean
  showSelectedOnlyOpen?: boolean
  showClearSelected?: boolean
  gridAlignment?: TextAlignValue

  showRefresh?: boolean

  // link
  linkParam?: IMLinkParam

  // search
  searchOpen?: boolean
  searchFields?: string[]
  searchExact?: boolean
  // filter
  filterOpen?: boolean
  filter?: SqlExpression
  // sort
  sortOpen?: boolean
  sorts?: SortSettingOption[]

  // card background
  cardConfigs?: ImmutableObject<{ [status: string]: CardConfig }>
  searchHint?: string

  isShowAutoRefresh?: boolean
  noDataMessage?: string
  layoutType?: ListLayoutType
  keepAspectRatio?: boolean
  gridItemSizeRatio?: number

  showRecordCount?: boolean
  hidePageTotal?: boolean
}

export type IMConfig = ImmutableObject<Config>